requirements.txt
langchain==0.3.0
langchain-openai==0.2.0
psycopg2-binary==2.9.9
pydantic==2.9.0
pypdf==5.0.0
python-dotenv==1.0.1